+++ uip/picksbr.c
@@ -538,7 +538,8 @@ PRaction (struct nexus *n, int level)
 		dasctime (&n->n_tws, TW_NULL));
 	return;
     }
-    fprintf (stderr, "UNKNOWN(0x%x)\n", (unsigned int) (*n->n_action));
+    fprintf (stderr, "UNKNOWN(0x%x)\n",
+	     (unsigned int)(unsigned long) (*n->n_action));
 }
